fix bugs in mapping

+++ RoamerEngine/src/roamer_engine/display/Texture2D.cpp
@@ -31,7 +31,7 @@ namespace qy::cg {
 		} else if (nrChannels == 4) {
 			glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, width, height, 0, GL_RGBA, GL_UNSIGNED_BYTE, data);
 		} else if (nrChannels == 1) {
-			glTexImage2D(GL_TEXTURE_2D, 0, GL_R16, width, height, 0, GL_RED, GL_UNSIGNED_SHORT, data);
+			glTexImage2D(GL_TEXTURE_2D, 0, GL_R8, width, height, 0, GL_RED, GL_UNSIGNED_SHORT, data);
 		} else {
 			throw std::runtime_error("Not supported image format!");
 		}
